# MethodBuilder optimizations