@zacharycmontoya 2020-06-23
This looks good! Do you need to update the integrations to consume the `WithParameters` changes or are the integrations going to keep their existing codepaths?