@kevingosse 2020-06-24
> This looks good! Do you need to update the integrations to consume the `WithParameters` changes or are the integrations going to keep their existing codepaths?

Generic overloads have a higher priority, so the integrations should automatically use them